requirements.txt
imageio==2.31.5
matplotlib==3.8.0
numpy==2.0.1
opencv_python==4.8.1.78
opencv_python_headless==4.10.0.84
pandas==2.2.2
patchify==0.2.3
Pillow==10.0.1
Pillow==10.4.0
scikit_learn==1.3.1
skimage==0.0
tifffile==2023.9.26
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
